requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Command==0.1.0
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.7
Django==5.0.2
djangorestframework==3.14.0
filelock==3.12.4
Flask==3.0.0
fonttools==4.44.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.1
packaging==23.2
panda==0.3.1
pandas==2.1.3
patsy==0.5.3
Pillow==10.1.0
platformdirs==3.11.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sklearn==0.0.post11
sqlparse==0.4.4
statsmodels==0.14.0
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
Werkzeug==3.0.1
